:sparkles: Added event announcing and organizer selection and first admin commands.

+++ src/main/kotlin/services/ConfigService.kt
@@ -48,7 +48,7 @@ class ConfigService {
 
 @Serializable
 data class Config(
-    val telegramBotToken: String, val telegramBotUsername: String, val debug: Boolean = false, val managingGroup: Long, val adminIds: List<Long> = emptyList(),
+    val telegramBotToken: String, val telegramBotUsername: String, val debug: Boolean = false, val managingGroup: Long, val adminIds: List<Int> = emptyList(),
     val cron: CronConfig = CronConfig(), val locationPoll: LocationPollConfig = LocationPollConfig(), val knownUsers: MutableMap<Int, String> = mutableMapOf()
 )
